+++ src-tauri/src/thunderstore/query.rs
@@ -236,7 +236,7 @@ where
     let search_term = args
         .search_term
         .as_ref()
-        .map(|s| s.to_lowercase().replace(' ', ""));
+        .map(|s| s.to_lowercase().replace(' ', "_"));
 
     let mut result = mods
         .filter(|queryable| {